requirements.txt
noise==1.2.2
numpy==1.19.4
Pillow==8.1.0
pygame==2.0.1
